api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.1.0
python-multipart>=0.0.9
pyyaml>=6.0.1
